rename settings dictionaries to dictionary list (folder and string resources)
