+++ app/src/main/res/values/strings.xml
@@ -161,23 +161,22 @@
     <string name="modify_dictionary_validation_lang_to_empty">Language to must be not empty</string>
 
 
-    <!--     TODO remove to dictionaryList -->
-    <!--  Settings Dictionaries -->
-    <string name="setting_dictionaries_active_dialog_title">This dictionary active</string>
-    <string name="setting_dictionaries_active_dialog_description">You mark the dictionary active for the languages you are currently learning.
+    <!--  Dictionary List -->
+    <string name="dictionary_list_active_dialog_title">This dictionary active</string>
+    <string name="dictionary_list_active_dialog_description">You mark the dictionary active for the languages you are currently learning.
 When you create new word or new list this dictionary will be prefilled. Words to daily exam will take from this dictionary</string>
-    <plurals name="setting_dictionaries_delete_dictionary_title">
+    <plurals name="dictionary_list_delete_dictionary_title">
         <item quantity="one">Are you sure you want to delete %s dictionary</item>
         <item quantity="other">Are you sure you want to delete %s dictionaries</item>
     </plurals>
-    <plurals name="setting_dictionaries_delete_dictionary_description">
+    <plurals name="dictionary_list_delete_dictionary_description">
         <item quantity="one">All words in dictionary will be deleted</item>
         <item quantity="other">All words in these dictionaries will be deleted</item>
     </plurals>
-    <string name="setting_dictionaries_screen_title">Dictionaries</string>
-    <string name="setting_dictionaries_screen_title_cd">Navigation button to Dictionaries screen</string>
-    <string name="setting_dictionaries_screen_remove_active_dictionary">You try to delete active dictionary, if you do it, exam reminders will stop until you select a new active dictionary</string>
-    <plurals name="setting_dictionaries_screen_dictionary_was_deleted">
+    <string name="dictionary_list_screen_title">Dictionaries</string>
+    <string name="dictionary_list_screen_title_cd">Navigation button to Dictionaries screen</string>
+    <string name="dictionary_list_screen_remove_active_dictionary">You try to delete active dictionary, if you do it, exam reminders will stop until you select a new active dictionary</string>
+    <plurals name="dictionary_list_screen_dictionary_was_deleted">
         <item quantity="one">Dictionary was deleted</item>
         <item quantity="other">Dictionaries was deleted</item>
     </plurals>
